fix(importCss.js): To return promise in case href was used already
That would help in case you have multiple lazy chunks combined in a one file. For example, let's say
you have 2 lazy components and you'd like to load them together in terms of network requests. For
that reason you'll specify the same webpackChunkName for both. But, since you have the same name
href here for both components that would mean that the first component would be loaded properly,
because promise would be resolved only when css is loaded but the second component could be rendered
on the page without css because href is the same and promise would be resolved immediately. And if
we return the same promise back then both components would be rendered only after css is completely
fetched.