feat(loadOption): Nested chunknames as major release
Reverted this from 1.5.2 in order to re-introduce as a breaking change with a major release version
bump. I also added more conventional looking tests

BREAKING CHANGE: Chunknames no longer output directories. If the chunkname was components/component,
then the output would be a directory called components with the component inside it. Keeping with
webpack 3 chunking conventions, chunk folders are represented in the filename with a hyphen as a
seperator: components-component.js. the side affect is flattened build outputs

re #45