fix: support multiple imports in one file
This avoids caching the result of `addDefault` and offloads the
responsibility of deduplicating imports to that helper function.
As a side-effect, it appears that the imports are reprocessed as their
internal identifiers are reassigned N times for N imports, so this isn't
an optimal solution, but the resulting code is at least correct.