fix(babel-plugin-universal-import): Prevent exposing filename always and move it to optional setting
By default, file was always a property generated and pointed to the source file including full path
to the file. This is an unnecessary exposure of information. Moved to a flag setting. Also tests
were depending on this and were specific to *nix folder/file resolution, making updates to this
plugin from a different system impossible.